requirements.txt
genshin
python-dotenv